requirements.txt
aioredis==1.3.1
aiosqlite==0.16.1
async-timeout==3.0.1
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
Deprecated==1.2.12
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
hiredis==2.0.0
idna==3.1
iso8601==0.1.14
makefun==1.11.3
passlib==1.7.4
pycparser==2.20
pydantic==1.8.1
PyJWT==2.0.1
pypika-tortoise==0.1.0
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
ujson==4.0.2
uvicorn==0.13.4
wrapt==1.12.1
